\name{G.Estimate}
\alias{G.Estimate}
\title{
Estimates the most likely number of gametophytes in an eDNA sample
}
\description{
The function compares a user supplied vector with number of different alleles observed (genotyped) for each locus with sampling distributions created from a reference population \code{\link{MakeAllNumberDist}}. A numerical maximum likelihood method is used to estimate the most likely number of haploid (or diploid) genotypes that might have generated the observed number of different alleles observed across loci.
}
\usage{
G.Estimate(Obs.N.all,All.Num.Dist,ML.curve=TRUE,ploidy=1,
					col="black",col.est="darkcyan",col.ci="darkcyan")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Obs.N.all}{
A vector with length equal to the number of locus in the reference population allele frequency file. The values are the total number of different (unique) alleles genotyped at each locus from an eDNA sample. Locus with missing data should be coded with NA.
}
\item{All.Num.Dist}{
The 3-D array produced by \code{\link{MakeAllNumberDist}}. This array contains the sample distributions with unique allele counts generated by sampling the reference population allele frequencies, with or without Fis correction.
}
\item{ML.curve}{
A logical to decide if a plot with the maximum-likelihood curve should be produced; defaults to TRUE.
}
\item{ploidy}{
If estimating the number of haploids genotypes set to 1, if diploids set to 2.
}
\item{col}{
The color of the ML-curve line.
}
\item{col.est}{
The color of the vertical line intersecting the ML-curve at the estimated number of gametophytes.
}
\item{col.ci}{
The color of the vertical lines intersecting the ML-curve at the 95 percent confidence interval.
}
}
\details{
The probability that a certain number of observed alleles \emph{n.obs.A} at locus \emph{l} was produced by  \emph{g} gametophytes is estimated by the relative frequency of \emph{n.obs.A} in the sampling distribution for that locus and \emph{g} gametophytes,  recorded in \strong{All.Num.Dist}. The maximum likelihood is found by summing the log probabilities for all loci. An approximate confidence interval is produced using the likelihood ratio approach; in brief,  the 95 percent interval includes all gametophytes values for which the log-likelihood function drops off by no more than 1.92 units.
}
\value{
A list with the following components:
\item{estimateM }{A vector with three elements, the most likely number of gametophytes (G.hat), and the lower and upper bounds of the 95 percent confidence interval.}
\item{LogML}{The log-likelihoods for all values of gametophytes from one to the maximum number available in the sampling distributions. Inf values are returned when \emph{n.obs.A} are impossible given a particular number of gametophytes.}
}
\references{
Liggan L, Alberto F (in prep)...
}
\author{
F. Alberto and L. Liggan.
}
\seealso{
\code{\link{MakeAllNumberDist}}
}
\keyword{ IO }
