\name{G.Estimate}
\alias{G.Estimate}
\title{
Estimates the most likely number of haploid or diploid genotypes in an eDNA sample using the observed counts of unique microsatellite alleles amplified for a set of markers.
}
\description{
The function compares a user supplied vector with counts of unique alleles observed (genotyped) for each locus with sampling distributions created from a reference population \code{\link{MakeAllNumberDist}}. A numerical maximum likelihood method is used to estimate the most likely number of haploid or diploid genotypes that might have generated the observed counts of unique alleles observed across loci.
}
\usage{
G.Estimate(Obs.N.all,All.Num.Dist,ML.curve=TRUE,ploidy=1,
					col="black",col.est="darkcyan",col.ci="darkcyan")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Obs.N.all}{
A vector with length equal to the number of locus in the reference population allele frequency file. The values are the counts of unique alleles genotyped at each locus from an eDNA sample. Locus with missing data should be coded with NA.
}
\item{All.Num.Dist}{
The 3-D array produced by \code{\link{MakeAllNumberDist}}. This array contains the sample distributions with unique allele counts generated by sampling the reference population allele frequencies, with or without Fis correction.
}
\item{ML.curve}{
A logical to decide if a plot with the maximum-likelihood curve should be produced; defaults to TRUE.
}
\item{ploidy}{
If estimating the number of haploids genotypes set to 1, if diploids set to 2.
}
\item{col}{
The color of the ML-curve line.
}
\item{col.est}{
The color of the vertical line intersecting the ML-curve at the estimated number of gametophytes.
}
\item{col.ci}{
The color of the vertical lines intersecting the ML-curve at the 95 percent confidence interval.
}
}
\details{
The probability \emph{P(OCUA | g)} that a given count of unique microstallite alleles \emph{OCUA} (the vector elements in function argument \strong{Obs.N.all}) is amplified at locus \emph{i} from an eDNA sample was produced by \emph{g} genotypes is estimated using a numerical maximum likelihod approach. The relative frequency of \emph{OCUA} in different sampling distributions of expected counts of unique allele (ECUA) is calculated for samples of size \emph{g}, from 1 to \emph{G}.  The multilocus \emph{P(OCUA | g)} is calculated by summing the log probabilities for all loci for each \emph{g} value. The maximum likelihood estimate is simply the maximum multilocus probability across the values of \emph{g}. The sampling distributions for each locus and \emph{g}  combination  are created by sampling with replacement samples of size \emph{g} alleles according to their frequencies in the reference population, with or without Fis correction,  and counting the number of unique alleles \emph{CUA} in each sample. The sampled values of \emph{CUA}, building each sampling distribution, are recorded in \strong{All.Num.Dist}. An approximate confidence interval is produced using the likelihood ratio approach; in brief,  the 95 percent interval includes all genotype values for which the log-likelihood function drops off by no more than 1.92 units.
If estimating diploids the function simply assumes that the number of diploid genotypes is the double of haploid ones. Thus, for example, the distribution for four diploid genotypes is the same as the one generated for samples of eight genes.
An eDNA sample might fail to amplify all the microsatellite alleles present. More concerning, for problematic samples, there might be heterogeneity across loci in the number of alleles that are amplified. This can result in observed allele counts across loci that do not follow the different allele richness of each loci in the reference population. For example, a marker with more alleles in the reference population might have fewer amplified alleles in the eDNA sample than a loci with fewer alleles in the reference population. This will result in a biased low estimate of the number of genotypes present and non-informative confidence intervals. We used a simple chi-square test to measure how the observed number of alleles amplified compares with the expected number of alleles, given the diversity of the different loci in the reference population and the total number of alleles amplified. 
The chi-square test measures the quality of the eDNA sample. It provides complementary information to the CI because a lower allele richness, number of loci used, or both can also result in wide CI for eDNA samples with high quality (i.e., without amplification problems).
}
\value{
A list with the following components:
\item{estimateM }{A vector with three elements, the most likely number of genotypes (G.hat), and the lower and upper bounds of the 95 percent confidence interval.}
\item{LogML}{The log-likelihoods for all values of genotypes from one to the maximum number available in the sampling distributions. Inf values are returned when \emph{n.obs.A} are impossible given a particular number of genotypes.}
\item{Amp.Qual}{The results from a Chi-square test comparing the number of observed alleles scored in an eDNA sample with the expected number of alleles, given loci diversity in the reference population and the total number of alleles scored. The statistic ((0-E)^2)/E is given for each locus to allow detection of which loci deviate more from expectations (i.e., the loci where some alleles might have not amplified); the sum of these values, the Chi-square statistic, is given as well as the p-value with degrees of freedom equal to the number of loci minus two.}
}
\references{
Liggan L, Alberto F (in prep)...
}
\author{
F. Alberto and L. Liggan.
}
\seealso{
\code{\link{MakeAllNumberDist}}
}
\keyword{ IO }
