\name{MakeAllNumberDist}
\alias{MakeAllNumberDist}
\title{
Generates distributions of total number of alleles that can be produced by different number of gametophytes.
}
\description{
By resampling from the allele frequency table for the reference population, this function generates distributions of total number of alleles, that can be produced by different number of gametophytes.
}
\usage{
MakeAllNumberDist(genotypes.df,ngametos=80,All.nchar=3,Fis=FALSE,Fis.treshold=0.1,
npermutations=5000) 
}
\arguments{
  \item{genotypes.df}{
The input file with the reference population microsatellite allele codes. This is a data.frame with one column per locus one row per individual with column names containing the locus names and without individual information (i.e. the first column is the information for the first loci). Use the same number of digits to code each allele. For example, 120126 is a possible value for a diploid coded with three digits per allele. At this stage, the reference data needs to be in diploid format. Missing data needs to be coded with 0, 00, 0000, or 000000..
}
 \item{ngametos}{
The method will generate distributions of total number of unique alleles that could have been generated by samples containing one to g number of haploid genotypes. Distributions are sampled from the reference allele frequency table. There are distributions for each combination of locus and number of gametophytes.
}
\item{All.nchar}{
How many digits are used in the reference population file to code each allele (can only be 1, 2, or 3).
}
\item{Fis}{
A logical indicating if Fis corrections should be applied when generating the sampling distributions. If yes, Genepop is used to calculate Qintra, the observed frequencies of identical pairs of genes in the reference population. Alleles are then sampled sequential with probability Qintra to sample the previously sampled allele and 1-Qintra to sample any other allele according to their allelic frequencies. Defaults to FALSE.
}
\item{Fis.treshold}{
When applying Fis correction it is recommended to only correct for Fis values above a certain threshold. The default is 0.1.
}
 \item{npermutations}{
How many sample values should each distribution contain; defaults to 5000.
}
}
\details{
The function creates distributions of total number of alleles that are likely to be produced for each loci for different numbers of gametophytes (from one to ngametos).
}
\value{
A three-dimensional array with dimensions ngametos, number of loci (in the reference population allele frequency table), and npermutations, containing values for the total number of alleles sampled. This object will be required by the function \code{\link{G.Estimate}} to estimate the most likely number of gametophytes present in the eDNA sample.
}
\references{
Liggan L, Alberto F (in prep)...
}
\author{
F. Alberto and L. Liggan.
}
\seealso{
\code{\link{G.Estimate}}
}
\keyword{ IO }
